+++ multiqc/plots/bargraph.py
@@ -120,6 +120,13 @@ def plot (data, cats = None, pconfig = None):
                 if 'name' not in cat[c]:
                     cats[idx][c]['name'] = c
 
+    # Allow user to overwrite a given category config for this plot
+    if 'id' in pconfig and pconfig['id'] and pconfig['id'] in config.custom_plot_config:
+        for k, v in config.custom_plot_config[pconfig['id']].items():
+            if k in cats[idx].keys():
+                for kk, vv in v.items():
+                    cats[idx][k][kk] = vv
+
     # Parse the data into a chart friendly format
     plotsamples = list()
     plotdata = list()